requirements.txt
# Core dependencies
py-clob-client>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0

# Utilities
aiohttp>=3.9.0
